Extract `sending-post-capture.page` to `/home` feature.
