+++ src/app/features/home/home-routing.module.ts
@@ -47,6 +47,13 @@ const routes: Routes = [
         m => m.SeriesPageModule
       ),
   },
+  {
+    path: 'sending-post-capture',
+    loadChildren: () =>
+      import('./sending-post-capture/sending-post-capture.module').then(
+        m => m.SendingPostCapturePageModule
+      ),
+  },
 ];
 
 @NgModule({
