+++ src/app/features/home/sending-post-capture/sending-post-capture.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { SharedModule } from '../../../../../shared/shared.module';
+import { SharedModule } from '../../../shared/shared.module';
 import { SendingPostCapturePageRoutingModule } from './sending-post-capture-routing.module';
 import { SendingPostCapturePage } from './sending-post-capture.page';
