+++ src/app/features/home/sending-post-capture/sending-post-capture.page.scss
